services/cart/requirements.txt
flask==2.3.3
flask-restful==0.3.10
flask-sqlalchemy==3.1.1
SQLAlchemy==2.0.23
marshmallow==3.20.1
python-dotenv==1.0.0
gunicorn==21.2.0